LayoutTests/media/video-mouse-focus-expected.txt

This tests that a mouse click event will not cause a media element to gain focus.

EXPECTED (document.activeElement.id != 'video') OK
END OF TEST
